Remove libandroid_support from 64 bits targets.
libandroid_support is not supported in 64 bits targets in NDK
version 17 and 18, so trying to copy and link against it is not
necessary.

The change removes the copying of libandroid_support.a into the
result directory (which doesn't seem to be used for anything) and
moves the -landroid_support flag out of build.sh and into the
setCrossEnvironment files for 32 bits platforms.