Use PLATFORMVER consistently in the paths.
Even if PLATFORMVER was allowed as a envvar and set with a default,
the value was not used in the paths pointing to the NDK.

NDK r18 removes some of the API levels pointed by default in the
files, so it would never have work for r18. Using PLATFORMVER, one
can specify an API level that exist in r18 (like android-21) and
compile correctly.

This will not work by itself to compile in r18. It needs also #10.