Generate the ICU data into a library.
The archive method will fail for Android installations, unless
the archive data is also moved to the right place, which is not
always possible with Android.

The library method will create a library, which can be copied
almost anywhere inside the app, and loaded from there.

The Swift build system doesn't seem to support moving the data
archive, and it looks like the data should have been in the
library all along, so it also makes sense to have it there.